al parecer toda la logica con productos ya esta

+++ src/main/java/com/togamma/heladeria/dto/request/almacen/ProductoRequestDTO.java
@@ -1,12 +1,11 @@
 package com.togamma.heladeria.dto.request.almacen;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public record ProductoRequestDTO(
     String nombre,
     Boolean seVende,
-    BigDecimal precioUnitarioVenta,
+    Double precioUnitarioVenta,
     String unidadBase,
     Long idCategoria,
     List<RecetaItemRequestDTO> receta
