+++ src/main/java/com/togamma/heladeria/dto/response/almacen/ProductoResponseDTO.java
@@ -1,6 +1,5 @@
 package com.togamma.heladeria.dto.response.almacen;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -9,7 +8,7 @@ public record ProductoResponseDTO (
     LocalDateTime updatedAt,
     String nombre,
     Boolean seVende,
-    BigDecimal precioUnitarioVenta,
+    Double precioUnitarioVenta,
     String unidadBase,
     String categoria,
     List<RecetaItemResponseDTO> receta
