+++ src/main/java/com/togamma/heladeria/model/almacen/Producto.java
@@ -1,6 +1,5 @@
 package com.togamma.heladeria.model.almacen;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import org.hibernate.annotations.SQLDelete;
@@ -27,7 +26,7 @@ public class Producto extends BaseEntity {
 
     private String nombre;
     private Boolean seVende;
-    private BigDecimal precioUnitarioVenta;
+    private Double precioUnitarioVenta;
     private String unidadBase;
 
     @ManyToOne(fetch = FetchType.LAZY)
@@ -40,5 +39,4 @@ public class Producto extends BaseEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "id_empresa")
     private Empresa empresa;
-
 }
